int64

+++ pkg/entity/email.go
@@ -3,8 +3,8 @@ package entity
 import "time"
 
 type Email struct {
-	ID      int       `json:"id"`
-	UserID  int       `json:"userID"`
+	ID      int64     `json:"id"`
+	UserID  int64     `json:"userID"`
 	Address string    `json:"address"`
 	Created time.Time `json:"created"`
 }
