+++ pkg/entity/user.go
@@ -3,7 +3,7 @@ package entity
 import "time"
 
 type User struct {
-	ID      int       `json:"id"`
+	ID      int64     `json:"id"`
 	Name    string    `json:"name"`
 	Created time.Time `json:"created"`
 	Updated time.Time `json:"updated"`
